+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- Created a class for extracting psf frame for single order spectral data. This clss combines `flux_coo`, `norm_flux_coo`, `univariate_psf_frame` and `bivariate_psf_frame` functions.
+- Changed the name of `psf_extract` to `optimal_extract` to avoid confusion.
+- Updated docstrings
+
 ## [0.2.0] - 2023-02-24
 
 ### Added